requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.4
amqp==5.1.1
anyascii==0.3.2
apparmor==3.1.6
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
Babel==2.12.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.1.0
blinker==1.5
boto3==1.26.151
botocore==1.29.151
Brotli==1.0.9
btrfsutil==6.3.3
cachetools==5.3.1
case==1.5.3
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colour==0.1.5
configobj==5.0.8
cryptography==41.0.2
cssselect2==0.7.0
Cython==3.0.0
dbus-python==1.3.2
decorator==5.1.1
Django==4.0.4
django-ipware==5.0.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
dpath==2.1.6
duplicity==1.2.3
email-validator==2.0.0.post2
et-xmlfile==1.1.0
ewmh==0.1.6
execnet==1.9.0
extras==1.0.0
factory-boy==3.2.1
Faker==18.10.1
fasteners==0.18
fastjsonschema==2.18.0
filelock==3.12.2
fixtures==4.1.0
flake8==6.0.0
Flask==2.2.5
flask-babel==3.1.0
Flask-BabelEx==0.9.4
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
furl==2.1.3
future==0.18.3
geographiclib==2.0
geoip2==4.7.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
gssapi==1.8.2
gufw==22.4.0
httplib2==0.22.0
idna==3.4
importlib-metadata==5.0.0
inflect==7.0.0
iniconfig==2.0.0
ipaddress==1.0.23
isort==5.12.0
itsdangerous==2.1.2
jaraco.context==4.3.0
jaraco.functools==3.8.0
jaraco.text==3.11.1
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lamb @ git+https://bitbucket.org/kovsyk/lamb-core.git@7f18a52acdd7d8053e6ba3f843ec5b246b54b178
layoutswitcherlib==0.8.36
lazy==1.5
lazy-object-proxy==1.9.0
ldap3==2.9
lensfun==0.3.4
LibAppArmor==3.1.6
libfdt==1.7.0
lit==15.0.7.dev0
lxml==4.9.2
Mako==1.2.4
manjaro-sdk==0.1
MarkupSafe==2.1.3
material-color-utilities-python==0.1.5
maxminddb==2.3.0
mccabe==0.7.0
more-itertools==10.0.0
multidict==6.0.4
mypy-extensions==1.0.0
netsnmp-python==1.0a1
nftables==0.1
nose==1.3.7
npyscreen==4.10.5
oauth2client==4.1.3
openpyxl==3.1.2
ordered-set==4.1.0
orderedmultidict==1.0.1
packaging==23.1
pacman-mirrors==4.23.2
paramiko==2.11.1
passlib==1.7.4
pathspec==0.11.1
pbr==5.11.1
phonenumbers==8.13.13
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.0.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
psycopg2==2.9.7
psycopg2-binary==2.9.6
pulsectl==23.5.2
pushjack==1.6.0
pwquality==1.4.5
py==1.11.0
pyaml==23.5.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.24.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
PyDrive2==1.16.1
pyflakes==3.0.1
Pygments==2.15.1
PyGObject==3.44.1
PyJWT==2.3.0
PyNaCl==1.4.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySide6==6.5.2
pytest==7.1.2
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-magic==0.4.27
python-redis-lock==4.0.0
python-xlib==0.33
pytz==2023.3
pyxdg==0.28
PyYAML==6.0.1
ranger-fm==1.9.3
redis==4.5.5
regex==2023.6.3
reportlab==3.6.12
requests==2.28.2
requests-file==1.5.1
rsa==4.9
s3transfer==0.6.1
setproctitle==1.3.2
shiboken2==5.15.10
shiboken6==6.5.2
shiboken6-generator==6.5.2
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
speaklater==1.3
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.3
sshtunnel==0.4.0
svglib==1.5.1
systemd-python==235
testtools==2.6.0
tinycss2==1.2.1
tldextract==3.4.4
tomli==2.0.1
trove-classifiers==2023.8.8
typing_extensions==4.7.1
ueberzug==18.2.1
ufw==0.36.2
uritemplate==4.1.1
urllib3==1.26.15
validate==5.0.8
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.2
WTForms==3.0.1
xlwt==1.3.0
Yapsy==1.12.2
yarl==1.9.2
zipp==3.16.1
